fix file size update on save, format date/time and file size

+++ futurenhs.api/FutureNHS.Api/DataAccess/Models/FileAndFolder/FileVersion.cs
@@ -6,7 +6,7 @@ namespace FutureNHS.Api.DataAccess.Models.FileAndFolder
     {
         public Guid Id { get; init; }
         public string Name { get; init; }
-        public int? Size { get; init; }
+        public long? Size { get; init; }
         public Shared.Properties LastUpdated  { get; init; }
         
     }
